görev<T> genişletme: görev<metin> çalışıyor (runtime i64 taşıma) + kesirli T reddi [D-294]
D-293 lambda tarafını çözmüştü (define ptr @lambda_0 ✓) ama kdl_gorev_birlestir
int32_t dönüyordu → görev<metin> LLVM tip hatası veriyordu. Runtime i64 taşımaya
geçti; birleştir sonucu çağrı yerinde T'ye daraltılıyor (ptr→inttoptr, i64→aynen,
i8/16/32/i1→trunc). görev<metin> artık ÇALIŞIYOR (uçtan uca exit 5).

T'nin IR'i BİLDİRİLEN görev<T>'den gelir (LlvmIsim.gorev_ic_ir) — görev<T> IR'de
opak ptr olduğu için T başka türlü bilinemez (D-293'teki kapanis_donus_ir deseni).

Neden i32 dönen lambda bozulmuyor: int64 çağrılınca üst 32 bit ÇÖP olur, codegen
sonucu T'ye TRUNC eder → değer doğru. Bu trunc kozmetik değil, ŞART.

KESİRLİ T REDDİ — KATMANLI SAVUNMA (ikisi de ölçüldü):
- Runtime sonucu TAMSAYI dönüşlü fn-ptr ile alır (x0/rax); kesirli dönüş v0/xmm0'da
  → bitcast SESSİZ çöp üretirdi. Reddetmek, hata modunu loud→silent'a çevirmemenin
  tek yolu (D-293'te tam bu tuzağa düşülüp yakalanmıştı).
- 1. katman (tip kontrolü): 2 tıkaç — görev_başlat (yaratma) + DUGUM_TIP_GOREV
  (annotasyon/parametre) → DRF001.
- 2. katman: `--llvm` tip kontrolünü ÇALIŞTIRMAZ (önceden var olan davranış; ÖLÇÜLDÜ:
  --check exit 1 iken --llvm exit 0). O yol da güvenli: T=double için emisyon
  `trunc i64 -> double` üretir, GEÇERSİZDİR → LLVM gürültülü reddeder ("invalid cast
  opcode"). Sessiz çöp yolu KAPALI.
- Kapasite kaybı YOK: görev<kesirli*> zaten hiç derlenmiyordu; kazanç düzgün tanı.

TEST TUZAKLARI (ikisi de yakalandı):
- İlk red-testim birleştir'i ÇAĞIRMIYORDU → geçersiz cast hiç emit edilmiyor, program
  derlenip geçiyordu → test YANLIŞ SEBEPLE kalıyordu. Çağıracak şekilde düzeltildi.
- test_gorev_rt.c hâlâ `extern int32_t kdl_gorev_birlestir` bildiriyordu; ayrı derleme
  birimi olduğu için linker uyuşmazlığı YAKALAMAZ ve test düşük 32 biti okuyup YANLIŞ
  SEBEPLE geçerdi → bildirimler tanımla birebir hizalandı (int64_t).

SINIR: görev<T> V1: T ∈ {≤64-bit tamsayı, işaretçi-benzeri (metin/&T/Dizi<T>), boş}.
kanal<T> ayrı ve daha dar (32-bit tamsayı, D-292) — sınırı runtime tamponundan.

KANIT: test_llvm 252/252 (+2), test_drf 49/49 (+3: D47/D48 kesirli reddi iki yolda,
D49 görev<metin> kabulü), test_gorev_rt 13/13 (ABI hizalı). Regresyon: codegen parite
76/76, SELF-HOST FIXPOINT ✓ (33371 satır), test_tumu "Tum testler gecti!".
Örnek gorev_temel.kem görev<metin> ile zenginleştirildi (exit 42).

YAN BULGU: CLAUDE.md'deki "LLVM v4 (dizi param/return, dizi length, generic islev
codegen)" maddesi ESKİMİŞ — hepsi ZATEN çalışıyor (ampirik: dizi param 42, dizi dönüş
42, dizi_boyut 3, generic 42, generic+metin 5). Roadmap güncellendi.

PARİTE BORCU (sürüyor): D-291→D-294 selfhost/codegen.kem'de YOK.

+++ runtime/kdl_runtime.c
@@ -558,11 +558,20 @@ __attribute__((noreturn)) void kdl_panik(const char *mesaj) {
  * fn-ptr->fn-ptr donusumu ise -Wcast-function-type uyarisi verir (ikisi de
  * olcüldü). Codegen ayni `ptr`yi her iki parametreye de gecer; her isaretci
  * YALNIZ kendi gercek tipiyle cagrildigi icin ne uyari ne UB olusur. */
-typedef int32_t (*KdlGorevBare)(void *rho);
-typedef int32_t (*KdlGorevKapanis)(void *rho, void *env);
+/* D-294: dönüş int32_t -> int64_t (T genişletme).
+ * Neden: `görev<metin>` gibi işaretçi taşıyan T'lerde i32 taşıma işaretçiyi
+ * kırpardı. i64 taşıma HER İKİ yönü de doğru kılar:
+ *   - i32 dönen lambda int64 olarak çağrılınca üst 32 bit ÇÖP olur; codegen
+ *     birleştir sonucunu T'ye TRUNC ettiği için değer doğru kalır.
+ *   - ptr dönen lambda -> tam işaretçi (x0/rax 64-bit).
+ * KESİRLİ T DESTEKLENMEZ (tip kontrolünde reddedilir): float dönüş v0/xmm0'da
+ * gelir, buradaki tamsayı-dönüşlü çağrı x0/rax okur -> SESSİZ çöp. Reddetmek
+ * yerine bitcast'lamak hata modunu loud->silent'a çevirirdi. */
+typedef int64_t (*KdlGorevBare)(void *rho);
+typedef int64_t (*KdlGorevKapanis)(void *rho, void *env);
 
 typedef struct {
-    int32_t result;
+    int64_t result;
     int done;
     int32_t (*f)(void);     /* gorev islevi (eski kdl_gorev_basla_i32 yolu) */
     /* Katman 2 yolu (kdl_gorev_basla_kapanis) — f==NULL iken kullanilir */
@@ -683,7 +692,7 @@ KdlGorev *kdl_gorev_basla_kapanis(KdlGorevBare fn_bare,
  * sizintisiyla ayni status quo. kdl_bolge_bakiye() bunu durustce raporlar.
  * Serbest birakma, ayri bir F4-sinifi is olarak ρ_sahip confinement kanitiyla
  * birlikte gelir. */
-int32_t kdl_gorev_birlestir(KdlGorev *g) {
+int64_t kdl_gorev_birlestir(KdlGorev *g) {
     if (!g) return 0;
 #ifdef KDL_THREAD_WIN
     if (g->handle) {
@@ -693,7 +702,7 @@ int32_t kdl_gorev_birlestir(KdlGorev *g) {
 #elif defined(KDL_THREAD_POSIX)
     if (g->thr_valid) pthread_join(g->thr, NULL);
 #endif
-    int32_t r = g->result;
+    int64_t r = g->result;   /* D-294: T'ye daraltma codegen'de (trunc/inttoptr) */
     /* g->rho BILEREK serbest edilmiyor (yukaridaki not). */
     free(g);
     return r;
